fix(collector): enable multi-node custom allreduce on GB300 (#1416)
Pass dtype into TRT-LLM AllReduce so MNNVL is constructed on multi-node
TP (without it the collector silently records the fallback path). Add
2/4-node slurm launchers and prefer a measured TP8/TP16 slice over the
node-capped extrapolation.

Includes GB300 TP8/TP16 silicon for sglang 0.5.14, trtllm 1.3.0rc20, and
vllm 0.27.0 (184 rows across TP 2/4/8/16, 127MiB-1GiB payloads, custom
allreduce eager+graph).

+++ collector/network/collect_all_reduce.py
@@ -169,7 +169,9 @@ def benchmark_trtllm_allreduce(
 
         op_list = []
         for i in range(repeat_n):
-            allreduce = trtllm_mods["AllReduce"](mapping=mapping).cuda()
+            # dtype enables MNNVL for multi-node TP (issue #1416):
+            # _torch/distributed/ops.py @v1.3.0rc20 builds `MNNVLAllReduce(mapping, dtype) if dtype else None`.
+            allreduce = trtllm_mods["AllReduce"](mapping=mapping, dtype=torch_dtype).cuda()
             allreduce(input_tensor, all_reduce_params=all_reduce_params)  # dry run to init
             op_list.append(allreduce)
