+++ collector/network/slurm/collect_allreduce.py
@@ -88,7 +88,9 @@ while min_size < max_size:
     # to reduce impact of L2 cache hit
     op_list = []
     for i in range(repeat_n):
-        allreduce = AllReduce(mapping=mapping).cuda()
+        # dtype enables MNNVL for multi-node TP (issue #1416):
+        # _torch/distributed/ops.py @v1.3.0rc20 builds `MNNVLAllReduce(mapping, dtype) if dtype else None`.
+        allreduce = AllReduce(mapping=mapping, dtype=torch.bfloat16).cuda()
         output = allreduce(input_tensor, all_reduce_params=all_reduce_params)  # dry run to init
         op_list.append(allreduce)
         # print(output)
